Update base for Update on "[pkg] add generic ZipFile Reader/Writer"
add a generic zip file reader/writer to torch.package in order to get rid of dependency on torch for non torchscript / tensor related usages of package. This also enables users to create a derived class from the zip file reader/writer classes to have their own serialization/deserialization if it's desired for performance needs.

Differential Revision: [D33970688](https://our.internmc.facebook.com/intern/diff/D33970688)

[ghstack-poisoned]

Update base for Update on "[pkg] add generic ZipFile Reader/Writer"

add a generic zip file reader/writer to torch.package in order to get rid of dependency on torch for non torchscript / tensor related usages of package. This also enables users to create a derived class from the zip file reader/writer classes to have their own serialization/deserialization if it's desired for performance needs.

Differential Revision: [D33970688](https://our.internmc.facebook.com/intern/diff/D33970688)

[ghstack-poisoned]

Update on "[pkg] add generic ZipFile Reader/Writer"

add a generic zip file reader/writer to torch.package in order to get rid of dependency on torch for non torchscript / tensor related usages of package. This also enables users to create a derived class from the zip file reader/writer classes to have their own serialization/deserialization if it's desired for performance needs.

Differential Revision: [D33970688](https://our.internmc.facebook.com/intern/diff/D33970688)

[ghstack-poisoned]
