+++ test/package/test_dependency_api.py
@@ -6,8 +6,19 @@ from sys import version_info
 from textwrap import dedent
 from unittest import skipIf
 
-from torch.package import EmptyMatchError, Importer, PackageExporter, PackageImporter
-from torch.package.package_exporter import PackagingError
+from torch.package import (
+    EmptyMatchError,
+    Importer,
+    PackageExporter,
+    PackageImporter,
+    PackagingError,
+)
+from torch.package.package_exporter_no_torch import (
+    PackageExporter as PackageExporterNoTorch,
+)
+from torch.package.package_importer_no_torch import (
+    PackageImporter as PackageImporterNoTorch,
+)
 from torch.testing._internal.common_utils import IS_WINDOWS, run_tests
 
 try:
@@ -24,13 +35,18 @@ class TestDependencyAPI(PackageTestCase):
     - deny()
     """
 
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.PackageImporter = PackageImporter
+        self.PackageExporter = PackageExporter
+
     def test_extern(self):
         buffer = BytesIO()
-        with PackageExporter(buffer) as he:
+        with self.PackageExporter(buffer) as he:
             he.extern(["package_a.subpackage", "module_a"])
             he.save_source_string("foo", "import package_a.subpackage; import module_a")
         buffer.seek(0)
-        hi = PackageImporter(buffer)
+        hi = self.PackageImporter(buffer)
         import module_a
         import package_a.subpackage
 
@@ -44,7 +60,7 @@ class TestDependencyAPI(PackageTestCase):
 
     def test_extern_glob(self):
         buffer = BytesIO()
-        with PackageExporter(buffer) as he:
+        with self.PackageExporter(buffer) as he:
             he.extern(["package_a.*", "module_*"])
             he.save_module("package_a")
             he.save_source_string(
@@ -57,7 +73,7 @@ class TestDependencyAPI(PackageTestCase):
                 ),
             )
         buffer.seek(0)
-        hi = PackageImporter(buffer)
+        hi = self.PackageImporter(buffer)
         import module_a
         import package_a.subpackage
 
@@ -78,7 +94,7 @@ class TestDependencyAPI(PackageTestCase):
 
         buffer = BytesIO()
         with self.assertRaisesRegex(EmptyMatchError, r"did not match any modules"):
-            with PackageExporter(buffer) as exporter:
+            with self.PackageExporter(buffer) as exporter:
                 exporter.extern(include=["package_b.*"], allow_empty=False)
                 exporter.save_module("package_a.subpackage")
 
@@ -89,7 +105,7 @@ class TestDependencyAPI(PackageTestCase):
         buffer = BytesIO()
 
         with self.assertRaisesRegex(PackagingError, "denied"):
-            with PackageExporter(buffer) as exporter:
+            with self.PackageExporter(buffer) as exporter:
                 exporter.deny(["package_a.subpackage", "module_a"])
                 exporter.save_source_string("foo", "import package_a.subpackage")
 
@@ -99,7 +115,7 @@ class TestDependencyAPI(PackageTestCase):
         """
         buffer = BytesIO()
         with self.assertRaises(PackagingError):
-            with PackageExporter(buffer) as exporter:
+            with self.PackageExporter(buffer) as exporter:
                 exporter.deny(["package_a.*", "module_*"])
                 exporter.save_source_string(
                     "test_module",
@@ -114,12 +130,12 @@ class TestDependencyAPI(PackageTestCase):
     @skipIf(version_info < (3, 7), "mock uses __getattr__ a 3.7 feature")
     def test_mock(self):
         buffer = BytesIO()
-        with PackageExporter(buffer) as he:
+        with self.PackageExporter(buffer) as he:
             he.mock(["package_a.subpackage", "module_a"])
             # Import something that dependso n package_a.subpackage
             he.save_source_string("foo", "import package_a.subpackage")
         buffer.seek(0)
-        hi = PackageImporter(buffer)
+        hi = self.PackageImporter(buffer)
         import package_a.subpackage
 
         _ = package_a.subpackage
@@ -135,7 +151,7 @@ class TestDependencyAPI(PackageTestCase):
     @skipIf(version_info < (3, 7), "mock uses __getattr__ a 3.7 feature")
     def test_mock_glob(self):
         buffer = BytesIO()
-        with PackageExporter(buffer) as he:
+        with self.PackageExporter(buffer) as he:
             he.mock(["package_a.*", "module*"])
             he.save_module("package_a")
             he.save_source_string(
@@ -148,7 +164,7 @@ class TestDependencyAPI(PackageTestCase):
                 ),
             )
         buffer.seek(0)
-        hi = PackageImporter(buffer)
+        hi = self.PackageImporter(buffer)
         import package_a.subpackage
 
         _ = package_a.subpackage
@@ -170,7 +186,7 @@ class TestDependencyAPI(PackageTestCase):
 
         buffer = BytesIO()
         with self.assertRaisesRegex(EmptyMatchError, r"did not match any modules"):
-            with PackageExporter(buffer) as exporter:
+            with self.PackageExporter(buffer) as exporter:
                 exporter.mock(include=["package_b.*"], allow_empty=False)
                 exporter.save_module("package_a.subpackage")
 
@@ -183,7 +199,7 @@ class TestDependencyAPI(PackageTestCase):
 
         buffer = BytesIO()
         with self.assertRaises(PackagingError):
-            with PackageExporter(buffer) as he:
+            with self.PackageExporter(buffer) as he:
                 he.mock(include="package_a.subpackage")
                 he.intern("**")
                 he.save_pickle("obj", "obj.pkl", obj2)
@@ -196,7 +212,7 @@ class TestDependencyAPI(PackageTestCase):
         obj2 = package_a.PackageAObject(obj)
 
         buffer = BytesIO()
-        with PackageExporter(buffer) as he:
+        with self.PackageExporter(buffer) as he:
             he.intern(include="package_a.**")
             he.mock("**")
             he.save_pickle("obj", "obj.pkl", obj2)
@@ -205,7 +221,7 @@ class TestDependencyAPI(PackageTestCase):
         """If an error occurs during packaging, it should not be shadowed by the allow_empty error."""
         buffer = BytesIO()
         with self.assertRaises(ModuleNotFoundError):
-            with PackageExporter(buffer) as pe:
+            with self.PackageExporter(buffer) as pe:
                 # Even though we did not extern a module that matches this
                 # pattern, we want to show the save_module error, not the allow_empty error.
 
@@ -222,7 +238,7 @@ class TestDependencyAPI(PackageTestCase):
         import package_a  # noqa: F401
 
         buffer = BytesIO()
-        with PackageExporter(buffer) as he:
+        with self.PackageExporter(buffer) as he:
             he.save_module("package_a")
 
     def test_intern_error(self):
@@ -235,7 +251,7 @@ class TestDependencyAPI(PackageTestCase):
         buffer = BytesIO()
 
         with self.assertRaises(PackagingError) as e:
-            with PackageExporter(buffer) as he:
+            with self.PackageExporter(buffer) as he:
                 he.save_pickle("obj", "obj.pkl", obj2)
 
         self.assertEqual(
@@ -250,7 +266,7 @@ class TestDependencyAPI(PackageTestCase):
         )
 
         # Interning all dependencies should work
-        with PackageExporter(buffer) as he:
+        with self.PackageExporter(buffer) as he:
             he.intern(["package_a", "package_a.subpackage"])
             he.save_pickle("obj", "obj.pkl", obj2)
 
@@ -281,7 +297,7 @@ class TestDependencyAPI(PackageTestCase):
         buffer = BytesIO()
 
         with self.assertRaises(PackagingError) as e:
-            with PackageExporter(buffer, importer=BrokenImporter()) as exporter:
+            with self.PackageExporter(buffer, importer=BrokenImporter()) as exporter:
                 exporter.intern(["foo", "bar"])
                 exporter.save_source_string("my_module", "import foo; import bar")
 
@@ -300,7 +316,7 @@ class TestDependencyAPI(PackageTestCase):
         """An incorrectly-formed import should raise a PackagingError."""
         buffer = BytesIO()
         with self.assertRaises(PackagingError) as e:
-            with PackageExporter(buffer) as exporter:
+            with self.PackageExporter(buffer) as exporter:
                 # This import will fail to load.
                 exporter.save_source_string("foo", "from ........ import lol")
 
@@ -319,12 +335,12 @@ class TestDependencyAPI(PackageTestCase):
     def test_repackage_mocked_module(self):
         """Re-packaging a package that contains a mocked module should work correctly."""
         buffer = BytesIO()
-        with PackageExporter(buffer) as exporter:
+        with self.PackageExporter(buffer) as exporter:
             exporter.mock("package_a")
             exporter.save_source_string("foo", "import package_a")
 
         buffer.seek(0)
-        importer = PackageImporter(buffer)
+        importer = self.PackageImporter(buffer)
         foo = importer.import_module("foo")
 
         # "package_a" should be mocked out.
@@ -334,18 +350,25 @@ class TestDependencyAPI(PackageTestCase):
         # Re-package the model, but intern the previously-mocked module and mock
         # everything else.
         buffer2 = BytesIO()
-        with PackageExporter(buffer2, importer=importer) as exporter:
+        with self.PackageExporter(buffer2, importer=importer) as exporter:
             exporter.intern("package_a")
             exporter.mock("**")
             exporter.save_source_string("foo", "import package_a")
 
         buffer2.seek(0)
-        importer2 = PackageImporter(buffer2)
+        importer2 = self.PackageImporter(buffer2)
         foo2 = importer2.import_module("foo")
 
         # "package_a" should still be mocked out.
         with self.assertRaises(NotImplementedError):
             foo2.package_a.get_something()
+
+
+class TestDependencyAPINoTorch(TestDependencyAPI):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.PackageImporter = PackageImporterNoTorch
+        self.PackageExporter = PackageExporterNoTorch
 
 
 if __name__ == "__main__":
