+++ test/package/test_dependency_hooks.py
@@ -5,6 +5,9 @@ from io import BytesIO
 from torch.package import (
     PackageExporter,
 )
+from torch.package.package_exporter_no_torch import (
+    PackageExporter as PackageExporterNoTorch,
+)
 from torch.testing._internal.common_utils import run_tests
 
 try:
@@ -20,6 +23,10 @@ class TestDependencyHooks(PackageTestCase):
     - register_extern_hook()
     """
 
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.PackageExporter = PackageExporter
+
     def test_single_hook(self):
         buffer = BytesIO()
 
@@ -28,7 +35,7 @@ class TestDependencyHooks(PackageTestCase):
         def my_extern_hook(package_exporter, module_name):
             my_externs.add(module_name)
 
-        with PackageExporter(buffer) as exporter:
+        with self.PackageExporter(buffer) as exporter:
             exporter.extern(["package_a.subpackage", "module_a"])
             exporter.register_extern_hook(my_extern_hook)
             exporter.save_source_string("foo", "import module_a")
@@ -47,7 +54,7 @@ class TestDependencyHooks(PackageTestCase):
         def my_extern_hook2(package_exporter, module_name):
             my_externs.remove(module_name)
 
-        with PackageExporter(buffer) as exporter:
+        with self.PackageExporter(buffer) as exporter:
             exporter.extern(["package_a.subpackage", "module_a"])
             exporter.register_extern_hook(my_extern_hook)
             exporter.register_extern_hook(my_extern_hook2)
@@ -67,7 +74,7 @@ class TestDependencyHooks(PackageTestCase):
         def my_mock_hook2(package_exporter, module_name):
             my_mocks.remove(module_name)
 
-        with PackageExporter(buffer) as exporter:
+        with self.PackageExporter(buffer) as exporter:
             exporter.mock(["package_a.subpackage", "module_a"])
             exporter.register_mock_hook(my_mock_hook)
             exporter.register_mock_hook(my_mock_hook2)
@@ -87,7 +94,7 @@ class TestDependencyHooks(PackageTestCase):
         def my_extern_hook2(package_exporter, module_name):
             my_externs2.add(module_name)
 
-        with PackageExporter(buffer) as exporter:
+        with self.PackageExporter(buffer) as exporter:
             exporter.extern(["package_a.subpackage", "module_a"])
             handle = exporter.register_extern_hook(my_extern_hook)
             exporter.register_extern_hook(my_extern_hook2)
@@ -109,7 +116,7 @@ class TestDependencyHooks(PackageTestCase):
         def my_mock_hook(package_exporter, module_name):
             my_mocks.add(module_name)
 
-        with PackageExporter(buffer) as exporter:
+        with self.PackageExporter(buffer) as exporter:
             exporter.extern("module_a")
             exporter.mock("package_a")
             exporter.register_extern_hook(my_extern_hook)
@@ -118,6 +125,12 @@ class TestDependencyHooks(PackageTestCase):
 
         self.assertEqual(my_externs, set(["module_a"]))
         self.assertEqual(my_mocks, set(["package_a"]))
+
+
+class TestDependencyHooksNoTorch(TestDependencyHooks):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.PackageExporter = PackageExporterNoTorch
 
 
 if __name__ == "__main__":
